run checks every output
