+++ miseq_tools/pooling.py
@@ -6,6 +6,8 @@ def pooling(samplesheet, quant_csv):
     num_reads = (pooled_reads(samples) * 1e6).astype(int)
     concs = pd.read_csv(quant_csv, index_col=0)["nM"]
     pools = _pools(num_reads, concs)
+    _check_samples_used_exactly_once(pools, set(num_reads.index))
+    _check_dilution(pools, num_reads, concs)
     for i, pool in enumerate(pools, 1):
         print(f"Pool {i}")
         # show water first, then descending order of volume
@@ -85,3 +87,25 @@ def _pools(num_reads: dict[str, int],
         })
 
     return pools
+
+def _check_dilution(pools: list[dict[str, float]], num_reads: pd.Series, concs: pd.Series):
+    volume_so_far = pd.Series(0, index=num_reads.index, dtype=float)
+    for pool in pools:
+        volume_pool = pd.Series(pool, index=num_reads.index.to_list() + ['Water', 'Prev Pool']).fillna(0)
+        ul_prev_pool = volume_pool.pop('Prev Pool')
+        if ul_prev_pool > 0:
+            volume_pool += volume_so_far * ul_prev_pool / volume_so_far.sum()
+        volume_so_far = volume_pool
+    concs_final = concs * volume_so_far / volume_so_far.sum()
+    concs_final.drop('Water', inplace=True)
+    fracs_final = concs_final / 4
+    pd.testing.assert_series_equal(fracs_final[num_reads.index], num_reads / num_reads.sum())
+
+def _check_samples_used_exactly_once(pools: list[dict[str, float]], all_samples: set[str]):
+    unused_samples = all_samples.copy()
+    for pool in pools:
+        samples_in_this_pool = set(pool.keys()) - {'Water', 'Prev Pool'}
+        assert samples_in_this_pool.issubset(all_samples), f"Some sample(s) in pool {pool} are not valid samples"
+        assert samples_in_this_pool.issubset(unused_samples), f"Some sample(s) in pool {pool} have already been added to a previous pool"
+        unused_samples -= samples_in_this_pool
+    assert len(unused_samples) == 0, f"Samples {','.join(unused_samples)} have not been added to any pool"
